+++ apps/indieweb/handlers.py
@@ -2,28 +2,14 @@ import logging
 
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from files.utils import extract_uuid_from_url
-from post.models import TPost
-from webmention.models import WebMentionResponse
+from webmention import models as webmention_models
 
-from .models import TWebmention
+from .application import webmentions
 
 logger = logging.getLogger(__name__)
 
 
-@receiver(post_save, sender=WebMentionResponse)
+@receiver(post_save, sender=webmention_models.WebMentionResponse)
 def create_t_webmention(sender, instance, created, raw, using, update_fields, **kwargs):
-    uuid = extract_uuid_from_url(instance.response_to)
-    try:
-        t_post = TPost.objects.get(uuid=uuid)
-    except TPost.DoesNotExist:
-        logger.info("Webmention received for invalid post %s", uuid)
-        return
 
-    try:
-        t_webmention = TWebmention.objects.get(t_webmention_response=instance, t_post=t_post)
-        t_webmention.approval_status = None
-        t_webmention.update_microformat_data()
-    except TWebmention.DoesNotExist:
-        t_webmention = TWebmention.instance_from_webmentionresponse(instance, t_post)
-    t_webmention.save()
+    webmentions.create_moderation_record_for_webmention(instance)
