+++ apps/indieweb/models/_t_webmention_send.py
@@ -1,0 +1,18 @@
+from core.models import TimestampModel
+from django.db import models
+
+
+class TWebmentionSend(TimestampModel):
+    t_post = models.ForeignKey("post.TPost", related_name="ref_t_webmention_send", on_delete=models.CASCADE)
+    target = models.URLField()
+    dt_sent = models.DateTimeField()
+    success = models.BooleanField()
+
+    class Meta:
+        db_table = "t_webmention_send"
+        unique_together = ("target", "t_post")
+        verbose_name = "Sent Webmention"
+        verbose_name_plural = "Sent Webmentions"
+
+    def __str__(self):
+        return self.target
