+++ apps/indieweb/queries/__init__.py
@@ -1,0 +1,3 @@
+from ._t_webmention import get_webmention
+
+__all__ = ("get_webmention",)
