+++ apps/post/queries/__init__.py
@@ -1,0 +1,3 @@
+from ._t_post import get_t_post_by_uuid
+
+__all__ = ("get_t_post_by_uuid",)
