+++ apps/post/queries/_t_post.py
@@ -1,0 +1,5 @@
+from .. import models
+
+
+def get_t_post_by_uuid(uuid: str) -> models.TPost:
+    return models.TPost.objects.get(uuid=uuid)
