+++ requirements.lock
@@ -2,7 +2,7 @@ asgiref==3.3.2
 Django==3.2.9
 pytz==2020.5
 sqlparse==0.4.1
-django-webmention==2.0.1
+django-webmention==2.0.2
 djangorestframework==3.12.2
 beautifulsoup4==4.9.3
 gunicorn==20.0.4
